GetCombinedUpdatesMrc: split code to methods, changed config file, Fixed bugs.
Chagned GetCombinedUpdatesMrc to use the VoyagerToSolrConfiguration for
more properties. Now it uses the config for all output directories, for
input files, and the tmp directory.

Fixed bug with writing MARC data to files where if the recno was over
maxrec each MARC record would be saved to an individual file.

Fixed bug where if last recno module maxrec = 0 an empty file MARC would
be created.

Refactored code and split large run() method into smaller methods with
the goal of clarifying the high level activities of the method.

Added comments.

Part of issue DISCOVERYACCESS-865