Fixed bugs with IndexDirectory.
Fixed rdfToSolr/gradle.build java incremental tasks which had the wrong
names.

Improved IndexDirectory so it keeps the results in a HashMap.

Fixed IncrementalBibFileToSolr so that it examines the results from
IndexDirectory to see if there are any errors in the results. If there
are the utility will exit with an error.

Improved the reporting about file URLs and Solr URLs in
IncrementalBibFileToSolr.