Don't identify holdings records as current if their bib records are suppressed.
Generally, if a bib record is suppressed, any holdings records associated with it should
also be suppressed. This is far from a guarantee, and there are nearly always a handful
of outliers - unsuppressed holdings with suppressed bibs. In these cases, by not adding
the holdings record to the current records list, we can avoid the system thinking that
it should be reflected in Solr.