Fix queries for comparing the list of bibs in Voyager to Solr
While every bib record has a modified date in the record, Voyager's DB
table for bibs may show the modified date as null if it hasn't been
changed. When testing for an absent value in a SQL left join, this makes
the date a poor column to test for a null value.