Add unsuppressed record check when retrieving from Voyager for updates
For quite a few reasons, we may identify a suppressed record as needing updating
in Solr. Previously, the step that converted the MARC XML to RDF filtered these
records out. We may still want to do that for full index builds, but for
incrementals we now are avoiding downing the records from Voyager at all. I set
this up to check the unsupressed status in the current records db rather than
directly in Voyager. The Voyager status is more authoritative, but may differ
from the status at the time that the record lists are pulled. Since that list is
the updates target, it's the one to use here.