Deal well with active bibs that lack holdings
When comparing the lists of active records in Voyager to those in Solr, we
should generally find that no active bib will lack at least one active
holding record. However, from time to time it can happen, and shouldn't be
taken as an error condition or allowed to break the code.