Include additional (bound-with) barcode data in various search modes
Now that we have barcodes from the holdings records for bound-with
works, we can include those in addition to the item record barcodes
in search. Because when an item contains multiple works someone may
need to find the primary record an item is associated with, the
original barcode data (sourced from the item records) will now be
boosted to ensure that it comes up first in relevance order.