Allow exact title boosting when searching with or without leading articles
title_exact was previously a copyField from title_t, which was populated
from RecordToDocumentMARC. I moved the responsibility for this field from
RecordToDocumentMARC to AuthorTitleResultSetToFields, where most the other
work with main title occurs. In the new location, the two fields will be
populated separately. By doing so, title_exact can be a multiValued field
(which title_t is not). This allows title_exact to be useful for records
with vernacular titles where it previously was doing no good. The new
title_exact may have up to four values for a single record:

* 245 full title
* 245 full title w/o leading article
* 880 245 full title
* 880 245 full title w/o leading article

That last permutation should be the least frequent, as much of our 880
data is in languages that don't use articles at all (Chinese, Japanese,
Korean, Thai), but some is in languages that do (Hebrew, Arabic). Other
languages I'm not sure of.

Because this change involves changes to the Solr schema, the code changes
will remain segregated in their side-branch until we're ready to fully
rebuild the index.