Create experimental model for RSTF debuging based on config files
If this works, we'll be able to identify specific ResultSetToFields
classes for debugging in a config file, and the SPARQLFieldMaker will
print any Solr fields generated by the specified RSTF classes to
stdout whenever these classes' toFields() methods are run.

This should be a generally helpful feature, but is in place in the
short term to test potentially dangerous modifications of
AuthorTitleResultSetToFields associated with DISCOVERYACCESS-2020.