Start to flesh out index contents tracking so we can detect updates for title linking
Title linking creates the requirement that we may need to update a record in Solr
that hasn't changed. If the record has an outgoing link to another record that has
changed or gone away, the link may need to be addressed. In order to appropriately
target these updates, we will need to begin tracking for each bib:

* What the label should look like for any record pointing to the bib
* When that label last changed
* When the bib was last updated in Solr

This is far from all in place, but hopefully with this update we can begin start by
initializing the tracking data.