Add mapping of timestamp field in Solr record to BibSolr<Core>.timestamp field in DB
In the long run, I think I may need to run this process in two different modes...
one where the timestamp column in the DB shows the actual Solr timestamps, and
one where it is allowed to default to all zeros. The latter case may be helpful
for situations where I'm initiating a set of database tables for a new index that
has yet to build. This mode will likely also require the ability to specify a target
Solr core and a source core for this step separately.