Automatically update database tracking contents of Solr Blacklight index
Poling Solr to regenerate this list takes about 100 minutes, most of
which is recreating data that hasn't changed. Instead, we want the
updates process to push the changes into the database as it pushes
the changes into Solr. It may be appropriate to disable this feature
during full index rebuilds for the sake of efficiency, running the
full 100 minute data generation once the initial full build is complete.