Add $3 to display for fields 360,300,5XX
DISCOVERYACCESS-2079
Many of the requested fields, including 264 and several 5XX already
had the $3 values in display. For the ones that already were
displaying the $3 they generally were also adding the value to
the search index. For the fields that were not already displaying $3,
I followed the request by adding the value only to display. If at
some point we want to take some $3 values out of the search, we
can do that.