Update population of call number facet values
DISCOVERYACCESS-1321
The call number sort field seems to be working okay, but I've moved the
population of the Solr field from the HoldingsAndItemsRSTF to the
CallNumberResultSetToFields, so that the bib call numbers can be used
when holdings call numbers are unavailable. I also attempted to block
non-LC call numbers from the sort regardless of whether they come from
the bib or holdings. That decision can be addressed later if it's not
what we want.

DISCOVERYACCESS-1780
I made an attempt to partially address this problem by blocking call
numbers from the facet if their initial alphabetic component is longer
than three letters. Call numbers like "Video 1234" should be addressed.