For title linking fields, use $s where present instead of $t for CTS.
The subfield $s in title linking fields (76X-78X) is a uniform title,
which is more reliable in search than the $t, which is a transcribed
title. Where both are present, attempting to use both in click-to-search
functionality would result in the least effective possible search, so a
cascading fallback to the $t where the $s is absent seems to be the best
option.