Organize references to table names in current records inventory DB
This had gotten more than a little messy. By hard coding the table name
we can ensure that all of the table names can be managed in a single
place. By tracing references to the new CurrentDBTable enum, we can find
all references to a table in the code so as to update in the case of a
change to the table structures.