Catch IOException getting display headings from Solr
This request has an extremely low rate of failure, but when a program
makes millions of requests, a program crash on IO error has been fairly
frequent. The new code catches the error, and waits 3 seconds before
trying again. This will happen indefinitely for now, so the program will
neither become a network nuisance (thanks to the delay) nor produce the
overt failure that might draw attention to an issue needing help. This
can be expanded on later if needed.