Stop processing for authority records we don't want
If the main heading in an authority record is not one of the heading
types we are using, for example an subdivision authority, we can skip
the record without further ado, rather than continue to process.

Along the way, this should also address a null pointer exception from
trying to derive a heading string without a heading type description
having been identified.