Add new search fields and qf/pf definitions
"Begins with" search requires a few kinds of left-anchored search modes.
We have search modes for general text fields, for CJK fields, and for
numeric fields. All of these still require testing, but here's version
1.

Text fields:

These fields match the analysis we use for general text searching, but
they apply a fixed prefix to each of the indexed and queried values. As
long as the query is executed as a phrase search, the presence of the
prefix requires that the two phrases both start in the same left-hand
position.

CJK fields:

For CJK fields, the new analysis removes all spaces from the string, and
employs ngram filtering against the entire value as a single token.

Numeric fields:

For numeric fields, there's no attempt to enforce that the search string
and the submitted value begin in the same place. Instead, the indexed
value is tokenized on spaces, and the "begins with" is enforced within
each token. The intent is that a value of 1234567 will be found with a
begins-with search for 123, even if the 1234567 is not the first thing
in the relevant field.