Populate titles into search w/o leading articles
Several MARC fields that contain titles also contain indications of how
many characters of those titles are initial article data that shouldn't
be used for sorting. For begins-with searching, we similarly want to
make sure that the titles without the articles are populated into search
fields so that begins-with searching will work whether or not a searcher
includes the article in their search.