Make Solr inventory keys for items/mfhds nonunique
DISCOVERYACCESS-2300
If the bib that an item has been reassigned to is indexed before the bib
is it unassigned from is re-indexed, then the inventory will temporarily
reflect the item twice, with two associations. Since this situation can't
be easily protected against, it's best that the important columns are
indexed with non-unique keys instead of unique ones.