Clean up handling of bibs being re-added
DISCOVERYACCESS-2352
When a bib is removed from Solr, it's entries in the bib Solr inventory
table and the bib2work mapping table are deactivated, rather than being
deleted. This is done to help ensure that other records are updated if
they had linked to the deleted bib. In the infrequent (but very real)
case of a bib being re-added to Solr after having been dropped, the
deactivated database rows need to be cleanly reactivated, and the record
treated as new, so that incoming links can be re-established.