Eliminate duplicate cross references
By not including the $f values from authority headings for works, we've
merged sets of similar headings in cases where they only differ in the
$f. When these merged authority records come with incoming cross
references that, also, differ only in the $f, then the cross references
come out as exactly the same. As an adjustment for this, we only want
one of any set of matching cross references. Because some cases may
legitimately require cross references that point between the same pairs
of headings, the only ones we want to eliminate are those that match
exactly.

For example:
Iowa Conservation Commission
-> see
-> see also Earlier Heading
Iowa State Conservation Commission

(apparently both of these pointers to the same place are desired, and
should be maintained.)

As a side note, we also want to begin ignoring the $f data on authority
430 fields were applicable.

DISCOVERYACCESS-2285