Remove alt forms from author/subject json
Also, restructure the subject json so it's a little easier to parse.
DISCOVERYACCESS-1525

Finally, drop the alt_forms database table from the headings database.
The alternate forms can be queried from the reference and headings
tables. This duplicates the fix applied for DISCOVERYACCESS-2285, but
will have the benefit of more thoroughly eliminating duplicates. It will
also have the side effect of showing the alternate forms on the info
view with the parenthetical additions. For example, the American
Compensation Association lists "ACA" as an alternate form, which will
now be listed in the info screen as "ACA (American Compensation
Association)" instead.