Instruct Solr to escape backslashes in CSV
It looks like the occasional error messages have been the result of Solr
not excaping backslashes that appear in titles. In those rare cases
where a title appears at the end of a title it is interpreted by the CSV
parser as escaping the comma that separates the title from the next
field, causing the rest of the fields to be misidentified. Instructing
Solr to escape the backslashes seems to produce more valid CSV output.

DISCOVERYACCESS-1291