Switch to Apache CSV library from OpenCSV
I might be able to get this to work with OpenCSV, but after a bunch of
trial and error, I'm still loosing commas from within field values.
Escaping them still results in OpenCSV removing the embedded commas. The
Apache tool looks a little clearer, though we'll have to see how it
shapes up.