Don't define keys until we are done with load
It looks like telling MySQL to disable keys doesn't really do as much as
we'd like to hope. The load process has a sharp downturn in load speed,
which I suspect is due to the increased complexity of adding new values
to a total of seven different key indexes. Instead, I want to try
building the tables without the keys being established at all. Adding
them after the fact should take a while, but writing the keys in one go
ought to be quicker than maintaining them.