Add subfields to notes display
From the Rare Materials Cataloging Coordinator:

Just a couple of small adjustments that we wondered whether you could
make in Blacklight: Would it be possible to turn on display of subfield
d in the 544 field (the subfield for the title of a related collection)?
Also, could you please check to see whether subfields a and c are
displaying in the 541 field? I’m seeing some strange results there, as
in the case of A preliminary discourse on the study of natural history,
by William Swainson (History of Science QH51.S93). In the 541 note
there, subfield 3 is displaying, but not the rest of the note.