Dequeue reindexing of deleted/suppressed recs
If a record is deleted or suppressed in the catalog, it will be deleted
from the OPAC index. If the current representation in the index is out
of date, this is no longer relevant. Should the record ever by
unsuppressed, it will be re-added using the current record and code at
that time.