Add /termsOfUse request handler
We already have the /database request handler, which serves the same
purpose of returning TOU information, but the request handler was
configured to serve only database because that was all of the TOU data
we were dealing with at the time. /termsOfUse does the same thing, but
without the restriction to return only database record information. At
some point, we may want to remove the the /database handler, but only if
it's no longer is use.