Fix the handling ofreindexes based on record deletes
The implementation of the new system for handling updates was incomplete
in this case, where the process that deletes the records was creating
lists of the bibs that should be re-indexed on that basis, but the new
process that pulls queued records for index was not referencing those
files. Now, these are being recorded according to the new system.

DISCOVERYACCESS-2482