Switch from SD.getFieldValue() to SDB.getFieldValues()
pullReferenceFields() needs to pull the data from both SolrDocument and
SolrInputDocument objects. It did depend on casting SID to SD in order
to use a single SD interface. The new Solr is disabling the casting in
favor of a shared SolrDocumentBase interface. Unfortunately, the
getFieldValue() method still has different behavior on these two
objects, so I'm switching to getFieldValues() which should return
consistent responses. Thankfully, the new behavior will simplify the
code a little, since it always returns a Collection<Object> rather than
an Object when only one value is available.

DISCOVERYACCESS-2458