Remove ‡h from fulltitle_display
And as a result, from the exact title boost field AND the title sort
field.

DISCOVERYACCESS-2020
Having the ‡h as part of the "exact" title means that the boost only
occurs if the user types in the title including the ‡h, which is almost
guaranteed not to occur unless someone is aware of the bug.