Don't squeeze main title and author data into a single search field
DISCOVERYACCESS-2131
bug preventing left anchored search of 245 titles without initial
articles that appear in the bib

I have no idea what the original point of making these search fields
single-valued may have been, but it is causing a bug with the ability to
ignore articles in left anchored main title fields due to the alternate
version of the title (without the article) being postpended to the
original version rather than submitted as a separate value. As a result,
the alternate form is not on the "left" of that value.