Let MySQL know we don't need entire index queue
We don't know exactly how many rows of the queue will be necessary to
find minUpdateBibCount unique bib ids, since duplicates are not
uncommon, but putting any limit at all on the query will allow MySQL to
know that it doesn't need to calculate the entire list.

Along the way, the variable that is no longer a constant shouldn't be
named in all caps.