Populate DataField.mainTag in SMFM
The query made by SMFM doesn't return a hasField relationship as '?p',
which allows for the automatic extraction of main tag information. (By
contrast, tag information may be 880, which doesn't describe the role of
the field.)

While it might be a good idea to refactor the SPARQL at some point, for
now it's sufficient to pass along the mainTag info to the method
creating the DataField objects.

DISCOVERYACCESS-2568