Exclude alternate forms from left-anchored search
The alternate forms sourced from the authority records can be
counterproductive because it can create the impression that the
left-anchored query has failed. For example, a left-anchored subject
search for "American" returned a record whose subject heading "National
characteristics, American" seemed to match a general subject search but
not the left-anchored one. The explanation was the included alternate
form, "American national characteristics."

DISCOVERYACCESS-2131