Don't block access links from availability when space in loccode
An entire set of online theses are not being flagged as "Online" because
their location codes are "serv,remo " rather than "serv,remo". The
trim() will not address the fact that the locations are not being mapped
correctly, but the online records will at least no longer have their
Online status suppressed.