Add extendedIndexing option to run until 11pm instead of 5pm
When we want to get more indexing done, we can keep running several
hours later into the evening, then skip the browse index updates that
day. The control flow is sloppy, and depends on the job to index
headings into MySQL failing so that the one to push them to Solr will
not run. A more sophisticated control flow is in order if all else goes
well.