Use data type that doesn't eliminate duplicates
The use of Sets instead of Lists to store the isCJK Booleans only caused
an issue with bib 3943144 because the 880 ‡6 linkages for the tables of
contents are miscoded in the record.  However, graceful handling of
miscoded 880 ‡6 links is the preferred behavior, as errors in the links
are not rare.

DISCOVERYACCESS-2637