Reflect age-in-solr records updates in index queue
The index queue is serving the role of providing history data about when
and why a record has been indexed or re-indexed in Solr. If the queue is
empty, records will be indexed because they are the least recently
updated records in the index, which ought to be reflected in the queue
so it will be part of the record later.