Remove suppressed/deleted bibs from update queue
If a bib is in the queue to be updated and is suppressed before that
happens, the  outstanding re-index request will be marked as complete.
If the record is queued for update while it is suppressed or deleted,
then that is an error and the queue entry can be deleted.