Apply time-of-day rules to IndexHeadings
I had originally tried to prevent the update to the browse indexes from
running simultaneously with the following day's updates to the
blacklight index by making sure that on the days the blacklight indexing
is extended until 11pm, the browse index updates will not be attempted
afterward. This is not working as well as hoped because the index
optimization task tends to run past midnight, so the browse index
updates are launched the next day and are not blocked because it is no
longer an extended indexing day.

Instead, IndexHeadings will have an optional blackout window as provided
by the execution command line arguments (or class constructor arg
array). If the job is run outside of the allowable hours of the day, it
will throw an IllegalStateException and quit. But, when the job is run
in other contexts than the daily index process, the blackout times do
not have to be configured and the job will run whenever needed. For
daily indexing, I'm setting a window of 2am-8pm as allowable times for
this task to launch and can adjust if circumstances change.

(I don't think IllegalStateException is a perfect fit to what is going
on here, but the javadocs read "Signals that a method has been invoked
at an illegal or inappropriate time." That's what's happening, even if I
think the intent is more along the lines of not launching jobs during
system shutdown, not querying values before initialization.