Adjust begins-with Solr analysis for advanced search
The search mode that we use for advanced search seems to default to
non-phrase search when the search strings look like they will only have
one token. However, the original begins with analysis always ends with
an additional token to what was entered, making the initial appearance
of a single token search incorrect. By using an analysis closer to the
analysis of the left-anchored call number search fields, we should get
around that by not adding tokens.