DISCOVERYACCESS-2492, DISCOVERYACCESS-2592
Both of these issues involved reorganizing the handling of subject
indexing, so they were hard to separate. Fields 700,710,711,600,610, and
611 should now all be separated into author and title data (where
applicable) based on the position of the subfield $t, and not the
subfield codes of subfields before or after the $t. For the sake of term
agreement, the authorities indexing has been modified to use the same
logic. All subject terms will be indexed for search and browse, so the
subject CTS issue should be resolved in the new index. 630 fields will
now be flagged as "work" rather than as "general heading", which is more
accurate, while many of the uncontrolled fields previously omitted from
browse will be flagged as "general heading". The new general heading
browse field (subject_gen_browse) has been added to the Solr
configuration for subject click-to-search.