Switch to KStem for stemming
The KStem stemmer offers a combination of pattern and dictionary-based
stemming for a higher accuracy than the purely pattern-based
SnowballPorter algorithm. It doesn't support a protected word list the
way that the Porter stemmer does, but none of the words we are currently
protecting will need to be protected under the KStem stemmer. Any future
words that are found to need protection can be otherwise protected,
likely using the Synonym filter.

DISCOVERYACCESS-1294