Supply Blacklight search fields with browse records
Currently, the browse front end code is inferring the blacklight search
field to send a query to based on the type of browse and the description
of the field type (e.g. subject browse for "Personal Name" =>
subject_pers_browse). This works well enough, but it puts responsibility
on the front end that it doesn't need and introduces an additional point
that needs to be coordinated carefully through changes in indexing. In
this instance, we will have a new "General Heading" value in the subject
browse, and rather than add that mapping to the front end controls this
is probably a good time to simplify the front end controls by providing
the search field with the heading.

DISCOVERYACCESS-2592