Move record queueing to a dedicated class
The new method is blocking, to ensure that in the case of parallel
processing the same records won't be queued by multiple processes. By
blocking the queue table, other processes may also be delayed, which
probably won't matter as this is a quick process. Still, a different
locking mechanism might be a good idea in the future, perhaps locking a
dummy table that we don't actually need to use.

It accepts both a min and a max record count, and will not supplement
the indexing queue with older Solr documents unless they are needed to
mean the min count. I didn't protect against odd behavior in the case
where the provided min is greater than the provided max, but in that
case at most maxCount records will be taken from the queue, and minCount
will be reached by supplementing with older Solr documents.