Add processing mode that simply works the queue
Based on the iterative daytime indexing job, I needed one that took
somwhat larger batches (may increase further), and did not query Voyager
for updates between processing batches.

Also added a new MARC download service that should work for Bib, Mfhd
and Authority MARC files, and has the option to return the record as a
String (MARC21 or MARCXML) rather than only the option to save the file
to WebDav. Once this is tested, it should allow GetCombinedUpdatesMrc to
be phased out.