Lock dummy DB table instead of real one.
I thought that I could lock the table I was going to work with and would
still be able to work with it from the same connection. Turns out not so
much. The table batchLock exists only to prevent a concurrent instance
from trying to generate an updates batch while the first one does so.