Try to prevent table lock error
I'm not sure what caused this error this time, but the execution of the
batch update failed due to an existing read lock on the queue table.  I
suspect it's because I was populating rows in the update query while the
select query was still open, so I'm going to try creating the batch
update query later.