Don't commit Solr index during ProcessQueue
The inclusion of the call to commit in the IncrementalBibFileToSolr
meant that all indexing modes that use that method would commit at the
end of each file. (That's all modes except the Hadoop process.) For
parallel processing that's too much committing, so the commit needs to
be more optional, allowing ProcessQueue, which is designed to run in
parallel, to choose not to commit.