Add update mode for inventory of Voyager data & various cleanup
Rather than dropping the local tables that reflect Voyager contents and
rebuilding them, I want to do a side by side read down the local and
remote tables, modifying the local inventory tables only when
differences are found, and queueing the necessary index work at the same
time.

THIS NEW CODE IS NOT DONE.

The new mode should help by :
* Significantly speeding up the morning processing time before we're
ready to start actual work.
* Preventing running ProcessQueue jobs from acting inappropriately when
they query the local inventory tables and find them wildly incorrect, as
is the case when a table has been dropped and a fraction of it restored
so far.