Treat mfhds of recently suppressed bibs as suppressed
When populating mfhds into our inventory table we ignore mfhds whose
linked bibs are suppressed. We need to enforce the rule from the other
end, but dropping mfhds from our inventory if their bibs become
suppressed. At worst, a change in the unsuppressed mfhd for a suppressed
bib could and occasionally did trigger a re-index of the suppressed bib
based on the modification of its holdings.