Pull follow-up query into main query
The query for the active status of the bib is being called about 8
million times, as nearly all of the rows in the result set c_rs are
matched by rows in the result set v_rs. Adding this check nearly doubled
the execution time of this job. I hope that pulling the retrieval of the
bib's active status into the main query will be less time consuiming
than the 8 million follow-up queries.