Reverse change to left-anchored search analysis
In commit 663dae782ead9d176058238b6b9fab438a08387b, the beginswith
prefix was modified to be attached to the first token of the word to
correct for a search problem in the advanced search syntax. Thanks to
adjustments made to the begins-with advanced search, that's no longer
necessary. Keeping the prefix detached from the first token prevents the
K-Stem stemmer from behaving anomalously toward the first word in a
begins-with search due to failing to find the amalgamated word in its
dictionaries. It also prevents the WordDelimiterFilter from muddying the
waters while trying to decide whether the joined word is one word or
two.