Update synonyms file for use of K-Stemmer
K-Stemmer doesn't support a protwords.txt like the Snowbal Porter filter
did, so words that need to be protected from the K-Stemmer will need to
be modified to a form that it will nto attempt to stem.

I don't know why K-Stemmer is reluctant to stem the word "stars" to
"star", so the synonyms filter can accomplish that.