Display and search 541 fields with blank 1st indicator
It's a faster to implement if I say that we want the 541 fields with
blank or 1 for the 1st indicator, rather than the initially planned
negative test for a 0. We don't have the infrastructure for negative
indicator tests, and no other values exist in this indicator (at least
as of June).

DISCOVERYACCESS-2911