Round up to whole # of characters when removing # of bytes from string
If the specified number of bytes to removed from the beginning of a
string would result in taking only part of a wide Unicode character, the
original behavior would be to throw an exception rather than remove any
characters from the title. It looks like the most common cause for this
issue in practice will be the use of a wide character that a cataloger
didn't count as multiple bytes when cataloging, such as the use of a
non-standard apostrophe in one of the many languages that includes
apostrophe characters in articles. The most reliable response to this
condition is therefore probably to round up the number of bytes stripped
to include all of that last character.

DISCOVERYACCESS-2972