Ensure linking to unit library name
DISCOVERYACCESS-2720
The utiliity loading the library patterns was casting them to lower
case, which I had missed, so the location names need to be cast to lower
case as well. Also, switched the pattern map from a HashMap to a
LinkedHashMap to ensure that the comparison with each value will happen
in the priority order that they appear in the text file. This will also
affect the online site identification.