Don't append "For" when dealing with first subfield ‡v
In my first attempt, the first instrument after the ‡v thought that it
wasn't the first entriy because forAppended was true. As a result, we
ended up with a semicolon after the word For. In this modification, the
word For will only be appended when something is added to follow after
it.

DISCOVERYACCESS-1608

Bad sample data:

[First work]: For; flute; English horn; keyboards (2); percussion (2);
violin; cello; digital audio signal processing (2). Total performers: 10
[Second work]: For; flute; alto flute; piccolo; oboe; English horn;
clarinet; bass clarinet; piano; percussion (2); violin; viola; cello;
digital audio signal processing (2). Total performers: 15
[Third work]: For; drum set; cello. Total performers: 2