Apply standardization of apostrophe characters more rigorously in search fields
DISCOVERYACCESS-1408
This is working now for fields that are generated in the
StandardMARCFieldMaker using VernMode.SEARCH, but has been only
haphazardly applied to instances of resultSetToFields methods. One of
those not applied is the main aithor/title RSTF. Rather than apply the
change to each RSTF code block, this moves the resonsibility to the
addField() utility they all use.