Handle various data with order-enforcing types
TreeSet and TreeMap both return their contents according to _comparable_
order, which helps ensure that even if the records are received in an
arbitrary order they will be added to the Solr document in a predictable
order.

This is only desirable behavior for fields that do not need to be
displayed, or which will be displayed but not together (e.g. facet
values, browse entries). When display in record order is important, then
LinkedHashSet or LinkedHashMap is the correct option to enforce
_insertion_ order.

I'm hoping that most of these non-deterministic cases will be repaired
by making DataField comparable and putting the FieldSet list of
DataField objects into an ordered TreeSet, ensuring that various RSTF
code blocks process them in a predictable order, as most RSTF classes
use FieldSets to connect fields with their 880 matches.

DISCOVERYACCESS-3071