Correct (again) the handling of non-filing character counts
DISCOVERYACCESS-2972
My original implemenation of this issue was based on the belief that the
non-filing character count was a byte count of the UTF-8 encoding of the
beginning of the title value. This was close to a correct answer, but
breaks down for non-Roman scrpts where it imagines that catalogers will
understand the width of their Unicode characters when encoding the
value.  The correct understanding is that each grapheme is treated as
one base character, plus one for every diacritic it uses.. The new
implementation casts the string to Unicode decomposed form so it can
reliably find and count the diacritics.