Enable keys on BIB_VOY table after populating
This is one of those cases where after finding the problem apparently
causing a new glitch, it's hard to explain why it ever worked correctly.

After pulling the current bib list from Voyager into Mysql in 27 minutes
(normal), the mfhd list pull started running very slowly so that after 8
hours of processing it was only about 1% populated. On examination of
the database files, the MYI file for the BIB_VOY table was too small to
contain a working index.