Query terms counts from each Solr shard separately, where applicable
Eitehr way, the blacklightSolrUrl value is still used for determination
of standard display forms for non-authority headings.

DISCOVERYACCESS-2458
This change is motivated more by the move to SolrCloud than the move to
Solr 6 specifically. Querying all shards together forces SolrCloud to
collate the results from all six shards. Combined with the large size of
the returned result set and the fact that Solr is optimized for
returning result pages of moderate size, this is a difficult request for
Solr and in trial runs so far as sometimes resulted in crashing one of
the Solr instances.