Optimize each shard before tabulating headings
It appears that running the optimize command against the whole
collection at once doesn't result in all of the shards being optimized.
I'm also moving the responsibility for running the index optimization
command from ConfirmSolrIndexCompleteness to IndexHeadings because:

* IH is already working with the list of individual shards
* The need for optimized index is due to the requirements of IH, so it's
a better fit.

From my testing it appears that there's no need for the optimize command
to be sent to the shard leader. All replicas will still end up optimized
and the leader will not change.

DISCOVERYACCESS-3133