Break facet count lists from Solr into paged chunks
Asking Solr to collate all facet results from very large facets seems to
be too large of an ask given the memory available to it, though it does
not seem to have struggled with the full tabulation of 3.35 million
personal author names and their record counts. Hopefully, breaking the
requests to Solr into million page chunks will resolve this, though for
later chunks, the identification of the previous N million headings will
still be needed to return set N+1.

I also removed the code relating to handling of lists of collection
shards, and the optimizeIndex call. This means that the current code has
no explicit calls to optimize the index, which should be technically
unnecessary.

DISCOVERYACCESS-3133