Fix some cases where 600-611/700-711 are displayed poorly
The attempt to fix the display of these fields per DISCOVERYACCESS-2492
worked in a lot of cases, but failed in two specific cases, first the ‡d
was not added to display for most X10 and X11 fields, which created a
very off-putting display when the ‡d contained necessary display
constants. Second, by using the presence of a ‡t as the sole indicator
that a field contains a work entry, a fair number of (not exclusively
older) fields were not identified as containing works because the
relevant data was in a ‡k. Some examples from different records:

610 2 0 ‡a Freemasons. ‡b Massachusetts. ‡k Constitutions.
700 0    ‡a Eulalia, ‡c Saint, of Mérida. ‡k Legend.
710 2    ‡a Catholic Church. ‡k Liturgy and ritual. Sacramentary.