Update BatchLogic for main iterative updater & clean up BatchLogic generally
The BatchLogic interface now has an explicit startNewBatch() method that
should be called at the before preparing each batch, rather than the
implicit assumption that BatchLogic implementations can maintain their
own state by assuming that targetBatchSize() will be called at the
beginning of each batching process.

The BatchLogic implementation of the IterativeupdateFromVoyager process
will now treat lower priority items in the index queue the same way it
treats unqueued "Age of Record in Solr" updates, allowing it to continue
processing normally in the face of code change update queueing and in
the future likely Age of Record in Solr pre-queueing.

The behavior of the BatchLogic implementation in ProcessQueue should be
unchanged.

Priority 1 updates are now established as secondary data change updates,
and is no longer an allowable priority for code change updates loaded
through QueueRecordsForSolrIndex. It is currently only used for Title
Link Updates, but may be used in the future for Voyager data changes
such as Hathi updates, etc..